+++ packages/shared/src/index.ts
@@ -7,3 +7,4 @@ export * from "./upload";
 export * from "./home";
 export * from "./mail";
 export * from "./backups";
+export * from "./settings";